+++ src/video/canvas/canvas_renderer.js
@@ -80,17 +80,35 @@ class CanvasRenderer extends Renderer {
     }
 
     /**
-     * Set a blend mode for the given context
+     * set a blend mode for the given context. <br>
+     * Supported blend mode between Canvas and WebGL remderer : <br>
+     * - "normal" : this is the default mode and draws new content on top of the existing content <br>
+     * <img src="images/normal-blendmode.png" width="510"/> <br>
+     * - "multiply" : the pixels of the top layer are multiplied with the corresponding pixel of the bottom layer. A darker picture is the result. <br>
+     * <img src="images/multiply-blendmode.png" width="510"/> <br>
+     * - "lighter" : where both content overlap the color is determined by adding color values. <br>
+     * <img src="images/lighter-blendmode.png" width="510"/> <br>
+     * - "screen" : The pixels are inverted, multiplied, and inverted again. A lighter picture is the result (opposite of multiply) <br>
+     * <img src="images/screen-blendmode.png" width="510"/> <br>
      * @name setBlendMode
+     * @see https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/globalCompositeOperation
      * @memberof CanvasRenderer.prototype
      * @function
-     * @param {string} [mode="normal"] blend mode : "normal", "multiply"
+     * @param {string} [mode="normal"] blend mode : "normal", "multiply", "lighter, "screen"
      * @param {CanvasRenderingContext2D} [context]
      */
-    setBlendMode(mode, context) {
+    setBlendMode(mode = "normal", context) {
         context = context || this.getContext();
         this.currentBlendMode = mode;
         switch (mode) {
+            case "screen" :
+                context.globalCompositeOperation = "screen";
+                break;
+
+            case "lighter" :
+                context.globalCompositeOperation = "lighter";
+                break;
+
             case "multiply" :
                 context.globalCompositeOperation = "multiply";
                 break;